requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
async-timeout==4.0.3
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
boto3==1.37.27
botocore==1.37.27
Brlapi==0.8.3
Brotli==1.0.9
catkin-pkg-modules==1.0.0
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debmake==4.3.2
decorator==4.4.2
defer==1.0.6
devscripts===2.22.1ubuntu1
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
duplicity==0.8.21
empy==3.3.4
exceptiongroup==1.2.2
faiss-gpu==1.7.2
fasteners==0.14.1
filelock==3.13.1
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
fsspec==2025.3.2
future==0.18.2
gast==0.5.2
gpg==1.16.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keyring==23.5.0
kiwisolver==1.3.2
langchain==0.3.22
langchain-core==0.3.50
langchain-text-splitters==0.3.7
langsmith==0.3.24
language-selector==0.1
lark==1.1.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
lorem==0.1.1
lorem-text==2.1
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.5.1
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
nltk==3.9.1
numpy==1.21.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
orjson==3.10.16
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
protobuf==6.30.2
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycocotools==2.0.8
pycodestyle==2.8.0
pycups==2.0.1
pydantic==2.11.2
pydantic_core==2.33.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pytest==6.2.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-ldap==3.2.0
python-magic==0.4.24
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2024.11.6
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==1.0.0
roman==3.3
rosdistro-modules==1.0.1
rospkg-modules==1.5.1
s3transfer==0.11.4
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.3.1
SQLAlchemy==2.0.40
ssh-import-id==5.11
sympy==1.13.1
systemd-python==234
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unidiff==0.5.5
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
zstandard==0.23.0
